requirements.txt
attrs==19.3.0
cupy-cuda102==7.7.0
fastrlock==0.5
iniconfig==1.0.1
llvmlite==0.34.0
more-itertools==8.4.0
numba==0.51.0
numpy==1.19.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
six==1.15.0
toml==0.10.1
